refactor: extract WebpackCLIPlugin
